Updated repository classes to access and modify data from database

+++ StoreApp/StoreApp.Library/IProduct.cs
@@ -10,7 +10,7 @@ namespace StoreApp.Library
 
         string ProductName { get; }
 
-        double Price { get; }
+        decimal Price { get; }
 
         void UpdatePrice(int newPrice);
     }
