+++ StoreApp/StoreApp.Library/Product.cs
@@ -9,10 +9,10 @@ namespace StoreApp.Library
     {
         private readonly int _productId = 0;
         private readonly string _productName = "";
-        private double _price = 0.0;
+        private decimal _price = 0.0M;
 
 
-        public Product(int id, string name, double price)
+        public Product(int id, string name, decimal price)
         {
             _productId = id;
             _productName = name;
@@ -24,7 +24,7 @@ namespace StoreApp.Library
         public string ProductName => _productName;
 
 
-        public double Price => _price;
+        public decimal Price => _price;
 
         public void UpdatePrice(int newPrice)
         {
